requirements.txt
absl-py==2.3.1
addict==2.4.0
aiofiles==23.2.1
altair==5.4.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.2
asttokens==3.0.1
attrs==25.3.0
backcall==0.2.0
black==24.8.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
cloudpickle==3.1.1
comm==0.2.3
ConfigArgParse==1.7.1
contourpy==1.1.1
cycler==0.12.1
Cython==3.1.3
dash==3.3.0
decorator==5.2.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@d38d7161247e64276d4c44d9d0605291e80bd969
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.121.3
fastjsonschema==2.21.2
ffmpy==0.5.0
filelock==3.13.1
Flask==3.0.3
fonttools==4.57.0
fsspec==2024.6.1
fvcore==0.1.5.post20221221
google-auth==2.40.3
google-auth-oauthlib==1.0.0
gradio==4.29.0
gradio_client==0.16.1
gradio_imageslider==0.0.18
grpcio==1.70.0
h11==0.16.0
h5py==3.11.0
hf-xet==1.1.8
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
hydra-core==1.3.2
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iopath==0.1.9
ipython==8.12.3
ipywidgets==8.1.8
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_core==5.8.1
jupyterlab_widgets==3.0.16
kiwisolver==1.4.7
lazy_loader==0.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
MultiScaleDeformableAttention==1.0
mypy_extensions==1.1.0
narwhals==1.42.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.0
numpy==1.24.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
omegaconf==2.3.0
open3d==0.19.0
opencv-python==4.12.0.88
orjson==3.10.15
packaging==25.0
pandas==2.0.3
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==6.5.0
portalocker==3.0.0
prompt_toolkit==3.0.52
protobuf==5.29.5
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.7
pycolmap==3.12.5
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.1.4
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
retrying==1.4.2
rich==14.2.0
rpds-py==0.20.1
rsa==4.9.1
ruff==0.14.6
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
semantic-version==2.10.0
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.44.0
submitit==1.5.3
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.19
tokenizers==0.20.3
tomli==2.2.1
tomlkit==0.12.0
torch==2.3.1+cu121
torchvision==0.18.1+cu121
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
triton==2.3.1
typer==0.20.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
uvicorn==0.33.0
wcwidth==0.2.14
websockets==11.0.3
Werkzeug==3.0.6
widgetsnbextension==4.0.15
yacs==0.1.8
zipp==3.20.2
